requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
python-dotenv
sentence-transformers
faiss-cpu
langchain-huggingface